Adds email campaign feature for voting
Implements functionality to send voting start and reminder emails to newsletter subscribers.

Introduces an admin interface to send these emails, including options to target all subscribers or a specific email address.

Also adds statistics to the admin dashboard to track the number of total, voted, and not voted subscribers.

Updates the number of votes allowed per group to 8.